requirements.txt
torch
timm
opencv-python
numpy
pyyaml
matplotlib
tqdm
trimesh
